Changed BigInteger to Long, to improve the server performance.
git-svn-id: https://svn.apache.org/repos/asf/directory/shared/trunk@517453 13f79535-47bb-0310-9956-ffa450edef68

+++ ldap/src/main/java/org/apache/directory/shared/ldap/filter/AssertionNode.java
@@ -20,9 +20,6 @@
 package org.apache.directory.shared.ldap.filter;
 
 
-import java.math.BigInteger;
-
-
 /**
  * Node used for the application of arbitrary predicates on return candidates.
  * Applies dynamic and programatic criteria for the selection of candidates for
@@ -36,7 +33,7 @@ import java.math.BigInteger;
 public class AssertionNode extends AbstractExprNode
 {
     /** Setting scan count to max */
-    private static final BigInteger MAX = BigInteger.valueOf( Integer.MAX_VALUE );
+    private static final Long MAX = Long.MAX_VALUE;
 
     /** Setting display string to [MAX] */
     private static final String MAX_STR = "[" + MAX.toString() + "]";
